requirements.txt
kiteconnect>=4.0.0
pandas>=1.5.0
numpy>=1.20.0
pytz>=2022.1
selenium>=4.0.0
webdriver-manager>=4.0.0
pyotp>=2.8.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
smartapi-python>=1.3.0
